Refactor Signal Handling for SignalManager (#42)
* Refactor Signal Handling for SignalManager

* Refactor signal handler init into main method

* Troubleshooting test failures

* Add IPC_DIR to signal calls in neon_audio

* Cleanup signal manager init

* Refactor create_signal and check_for_signal imports

* Update neon-utils dependency spec

* Resolve unit test errors

* Cleanup imports
Remove class-level config init

+++ neon_audio/tts/__init__.py
@@ -39,8 +39,9 @@ from mycroft_bus_client import Message
 from ovos_plugin_manager.tts import load_tts_plugin
 from neon_utils.logger import LOG
 from neon_utils.metrics_utils import Stopwatch
-from neon_utils import create_signal, check_for_signal
 from ovos_utils import resolve_resource_file
+
+from neon_utils.signal_utils import create_signal, check_for_signal, init_signal_bus
 
 try:
     from neon_core.language import DetectorFactory, TranslatorFactory
@@ -54,7 +55,16 @@ from mycroft.util import play_wav, play_mp3, get_cache_directory, curate_cache
 _TTS_ENV = deepcopy(os.environ)
 _TTS_ENV['PULSE_PROP'] = 'media.role=phone'
 
+_IPC_DIR = None
+
 EMPTY_PLAYBACK_QUEUE_TUPLE = (None, None, None, None, None)
+
+
+def _get_ipc_dir():
+    global _IPC_DIR
+    if not _IPC_DIR:
+        _IPC_DIR = get_neon_local_config()["dirVars"]["ipcDir"]
+    return _IPC_DIR
 
 
 class PlaybackThread(Thread):
@@ -85,7 +95,8 @@ class PlaybackThread(Thread):
         while not self.queue.empty():
             self.queue.get()
         try:
-            self.p.terminate()
+            if self.p:
+                self.p.terminate()
         except Exception as e:
             LOG.error(e)
 
@@ -278,6 +289,7 @@ class TTS(metaclass=ABCMeta):
             bus:    Mycroft messagebus connection
         """
         self.bus = bus
+        init_signal_bus(self.bus)
         self.playback.init(self)
 
     def get_tts(self, sentence, wav_file, request=None):
@@ -547,7 +559,7 @@ class TTS(metaclass=ABCMeta):
                     for response in response_audio_files:
                         self.queue.put((self.audio_ext, str(response), vis, ident, listen))
                 else:
-                    check_for_signal("isSpeaking")
+                    check_for_signal("isSpeaking", config={"ipc_path": _get_ipc_dir()})
 
     def viseme(self, phonemes):
         """Create visemes from phonemes. Needs to be implemented for all
